refactor: rename $types.ts to types.ts
