+++ src/routes/daily-info/types.ts
